fixed notepad fixed addtask componente,modularizarion of all components and improvement in workflow
